+++ supabase/migrations/20240901174833_scholars.sql
@@ -21,8 +21,9 @@ create table public.scholars (
 create function isSteward() 
 returns boolean 
 language sql
+security definer set search_path to ''
 as $$
-    select (exists (select id from scholars where id = auth.uid() and steward));
+    select (exists (select id from public.scholars where id = (select auth.uid()) and steward));
 $$;
 
 alter table public.scholars
@@ -35,15 +36,18 @@ create policy "Scholar metadata is public" on public.scholars
   for select to anon, authenticated using (true);
 
 create policy "Scholars can be edited by stewards and selves" on public.scholars
-  for update to anon, authenticated using (id = auth.uid() or isSteward());
+  for update to anon, authenticated using (id = (select auth.uid()) or isSteward());
 
 create policy "Scholars can remove themselves" on public.scholars
-  for delete to anon, authenticated using (id = auth.uid());
+  for delete to anon, authenticated using (id = (select auth.uid()));
 
 
 -- Take the new user and insert a row in scholars
 create function public.handle_new_scholar()
-returns trigger as $$
+returns trigger 
+security definer set search_path to ''
+language plpgsql
+as $$
 begin
   -- Insert the new user into the scholars table.
   insert into public.scholars (id, email)
@@ -51,7 +55,7 @@ begin
   -- Return the new row.
   return new;
 end;
-$$ language plpgsql security definer;
+$$;
 
 -- This trigger automatically creates a profile entry when a new user signs up via Supabase Auth.
 create trigger on_auth_user_created