+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 ### Fixed
 
 - Typo in editor compensation.
+- Resolved all `search_path` security warnings.
+- More efficent calls to `auth.uid()` in RLS policies.
 
 ### Maintenance